fix nested column handling of null and "null"
