+++ processing/src/test/java/org/apache/druid/segment/TestHelper.java
@@ -449,13 +449,20 @@ public class TestHelper
     }
   }
 
-  public static Map<String, Object> createExpectedMap(Object... vals)
+  public static Map<String, Object> makeMap(Object... vals)
+  {
+    return makeMap(true, vals);
+  }
+
+  public static Map<String, Object> makeMap(boolean explicitNulls, Object... vals)
   {
     Preconditions.checkArgument(vals.length % 2 == 0);
 
     Map<String, Object> theVals = new HashMap<>();
     for (int i = 0; i < vals.length; i += 2) {
-      theVals.put(vals[i].toString(), vals[i + 1]);
+      if (explicitNulls || vals[i + 1] != null) {
+        theVals.put(vals[i].toString(), vals[i + 1]);
+      }
     }
     return theVals;
   }
